refactor(cisco_ios): consolidate `show line` parser to cover IOS-XE
The IOS-XE ``show line`` output differs from IOS only by an extra ``Line``
column in the table and an optional trailing ``Line(s) not in async mode
-or- with no hardware support`` block. Rather than maintain two near-
identical parsers, broaden the IOS sibling to absorb both differences:

- Header regex tolerates the optional ``Line`` column.
- Row regex captures an optional ``line`` group (IOS-XE only) and anchors
  ``typ`` to uppercase letters so IOS rows are not mis-parsed.
- ``line`` is added to the schema as ``NotRequired``.
- ``inactive_line_ranges`` is added to the result schema as ``NotRequired``
  and is only emitted when the IOS-XE trailing block is present.
- Stack ``@register(OS.CISCO_IOSXE, "show line")`` on the existing class.

The ``tests/parsers/iosxe/show_line/001_basic`` fixture is preserved so the
IOS-XE branch remains under test, alongside the existing IOS fixture.
